Make non-view/non-scalar signal data C contiguous
This ensures that assumptions about item size and strides are not
violated.

We can't do this to every initial_values because we want to keep signal
views as views and some parts of the code depend on scalars being 0d
NumPy arrays (a C contiguous array is always at least 1d).

Fixes #1184.