Raise exception on invalid gains.
This requires also to fix the calculation of gains and biases for
sigmoid neurons. Because the intercepts for sigmoid neurons is actually
not an intercept, but the inflection point, valid intercepts depend on
whether the max. firing rate is below or above the inflection point.
This commit also adjust the default refractory time constant of the
sigmoid neurons to be play well with the default intercepts < 1 and
default max. firing rates between 200 and 400Hz. Furthermore, this
required some slight adjustments to some tests.

Fixes #1212.