squash! fixup! Ignore failed move of cache index and warn.
Moves the message to cache.py, which should be more clear for
people looking at code.

The reason that the `NotAddedToNetworkWarning` puts the message in
the exception itself is because it's called in multiple places.
This warning only appears once, so we don't have to make the code
less readable in order to eliminate repetition.