squash! Add advanced connection documentation
Suggestion: Add references to examples how the pre ensemble can
influence the decoded function. The added references are things
that I know of because I wrote them, but feel free to replace or add
other references if you know any. Also feel free to not include them
if they don't seem to be a good fit (especially the second reference I
am a bit unsure.)