Fix copying of Parameter.
The keys of the WeakKeyIDDictionaries do not get updated on a normal
copy operation. This is rarely a problem because Parameters are usually
class variables that will not get copied, but there is the possibility
to define instance parameters (slightly obscure usage, but is for
example used in the current SPA system). In this case the Parameter
instance needs to be copied. Adding the __getstate__ and __setstate__
handler allows to convert the WeakKeyIDDictionaries to normal
dictionaries with strong references and convert them back. This will
cause the keys to be updated (partially due to Pythons copy magic).

It also seems to me that obtaining strong references during the copy
process is a good idea because than none of the objects can disappear
during copying (or serialization).

Addresses #1266.