squash! fixup! moved ``normalize_encoders`` inside constructor
Change the order of arguments so that all ensemble arguments come
first and all arguments (label, seed) passed on to the super class
come last.

Technically, this is not backwards compatible if someone uses the
argument position instead of keywords to pass label and seed, but I
hope no one is doing this, because one would have to spell out all
the arguments and it is not very readable.